requirements.txt
briefconf
httpx
pymongo
feedgen
feedparser
schedule
pytz
beautifulsoup4
pytest-playwright
fastapi[all]
cachetools
SQLAlchemy
python-socketio[asyncio_client]
python-socketio[asgi]
aiohttp
